Use custom envs, refactor (#10)
* feat: add generic `Combinator` type

The `Combinator` standardizes how 2D arrays of conditions are understood
throughout the program. Items in the outer array are ORed together,
while items in inner arrays are ANDed together.

* feat: flesh out combinator with inspection functions

* feat: add environment detection features

* fix: properly move test file to mod.rs

* feat: implement EnvTrie creation and processing

* feat: add method to evaluate an EnvTrie

* refactor: get code back to a compilable state

* test: fix failing test

* refactor: evaluate config when building builder

`ConfigBuilder::build` is now where environments are evaluated and hoard
paths are determined. This means that the built `Config` should have all
of the information it needs to do its thing without extra processing.

* fix: resolve compiler warnings and errors

* chore: enable a bunch more checks

* fix: resolve compiler warnings and errors

* fix: all clippy errors

* doc: add all of the documentation