feat: enable env vars in pile paths #22 (#23)
* chore: bump to 0.2.0 because 0.1.0 exists from previous owner

* feat: expand env vars in pile paths

* test(env): add more tests, make env tests serial

Tests that involve setting and checking env variables have to be run in
serial or else they might corrupt each other's environments.

* fix: parsing variables inside other variable names

To simplify parsing, environment variables are not allowed to have `$`
in their names, as this may indicate an attempt to build a variable name
using other variable names. As variables are not expanded recursively,
the outer "wrapping" variable should be ignored.