fix: change how weighted map is calculated
Previously, a topological sort of a DAG gave every environment a unique
weight. When setting scores on nodes was fixed, this broke an
integration test that ensures an error occurs in ambiguous situations.

Given:

```
exclusivity = [
    ["foo", "bar"],
    ["baz", "bar"]
]

[hoards.example]
    "foo" = "/some/path"
    "baz" = "/other/path"
```

Should foo or baz be chosen? The answer is neither, because they are
equal precedence over bar. The DAG method gave them unique weights,
though, so no error occurred. This is fixed by simplifying how weights
are determined: just use the reverse-order index in the exclusivity
list.

In this case, inserting a mutually exclusive item between foo and bar
will cause foo to be chosen over baz. This is fine though, since we
interpret this to mean that foo is more specific than baz.