refactor: remove references to permissions
Permissions checks were among the last few things to get test coverage,
but there is not a reliable way to maintain permissions across
filesystems and/or operating systems, particularly when Windows is
involved.

This removes permissions support. In the future, there will be a change
to apply permissions on restore.